Updates based on comments
