+++ rnd/autogpt_builder/src/lib/utils.ts
@@ -194,3 +194,10 @@ export function getPrimaryCategoryColor(categories: Category[]): string {
   }
   return categoryColorMap[categories[0].category] || "bg-gray-300/[.7]";
 }
+
+export function filterBlocksByType<T>(
+  blocks: T[],
+  predicate: (block: T) => boolean
+): T[] {
+  return blocks.filter(predicate);
+}
